requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Trading APIs
ccxt>=4.0.0
alpaca-py>=0.8.0
akshare>=1.10.0
yfinance>=0.2.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Cache
redis>=4.5.0

# Logging
loguru>=0.7.0

# CLI
click>=8.1.0
rich>=13.0.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0
textual>=0.47.0
plotext>=5.2.8

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Async
aiohttp>=3.8.0
websockets>=11.0.0
